CLDR-14399 move SurveyAjax.JSONWriter to SurveyJSONWrapper
